backend/requirements.txt
fastapi
uvicorn
rpy2
psycopg2-binary
